+++ emg3d/fields.py
@@ -32,7 +32,7 @@ from scipy.constants import mu_0
 from emg3d import maps, models
 
 __all__ = ['Field', 'SourceField', 'get_source_field', 'get_receiver',
-           'get_h_field']
+           'get_receiver_response', 'get_h_field']
 
 
 class Field(np.ndarray):
@@ -50,11 +50,11 @@ class Field(np.ndarray):
     A `Field` can be initiated in three ways:
 
     1. ``Field(grid, dtype=complex)``:
-       Calling it with a :class:`TensorMesh` instance returns a
+       Calling it with a :class:`emg3d.meshes.TensorMesh` instance returns a
        `Field` instance of correct dimensions initiated with zeroes of data
        type `dtype`.
     2. ``Field(grid, field)``:
-       Calling it with a :class:`TensorMesh` instance and an
+       Calling it with a :class:`emg3d.meshes.TensorMesh` instance and an
        `ndarray` returns a `Field` instance of the provided `ndarray`, of same
        data type.
     3. ``Field(fx, fy, fz)``:
@@ -68,7 +68,7 @@ class Field(np.ndarray):
     Parameters
     ----------
 
-    fx_or_grid : :class:`TensorMesh` or ndarray
+    fx_or_grid : :class:`emg3d.meshes.TensorMesh` or ndarray
         Either a TensorMesh instance or an ndarray of shape grid.nEx or
         grid.vnEx. See explanations above. Only mandatory parameter; if the
         only one provided, it will initiate a zero-field of `dtype`.
@@ -134,25 +134,15 @@ class Field(np.ndarray):
                 print("* ERROR   :: Provided grid must be a 3D grid.")
                 raise ValueError
 
-        # Get Laplace parameter.
+        # Store frequency
         if freq is None and hasattr(fy_or_field, 'freq'):
             freq = fy_or_field._freq
         obj._freq = freq
-        if freq is not None:
-            if freq > 0:  # Frequency domain; s = iw = 2i*pi*f.
-                obj._sval = np.array(-2j*np.pi*freq)
-                obj._smu0 = np.array(-2j*np.pi*freq*mu_0)
-            elif freq < 0:  # Laplace domain; s.
-                obj._sval = np.array(freq)
-                obj._smu0 = np.array(freq*mu_0)
-            else:
-                print("* ERROR   :: `freq` must be >0 (frequency domain) "
-                      "or <0 (Laplace domain)."
-                      f"             Provided frequency: {freq} Hz.")
-                raise ValueError("Source error")
-        else:
-            obj._sval = None
-            obj._smu0 = None
+        if freq == 0.0:
+            print("* ERROR   :: `freq` must be >0 (frequency domain) "
+                  "or <0 (Laplace domain).\n"
+                  f"             Provided frequency: {freq} Hz.")
+            raise ValueError("Frequency error")
 
         return obj
 
@@ -168,8 +158,6 @@ class Field(np.ndarray):
         self.vnEy = getattr(obj, 'vnEy', None)
         self.vnEz = getattr(obj, 'vnEz', None)
         self._freq = getattr(obj, '_freq', None)
-        self._sval = getattr(obj, '_sval', None)
-        self._smu0 = getattr(obj, '_smu0', None)
 
     def __reduce__(self):
         """Customize __reduce__ to make `Field` work with pickle.
@@ -180,8 +168,7 @@ class Field(np.ndarray):
 
         # Create our own tuple to pass to __setstate__.
         new_state = pickled_state[2]
-        attr_list = ['nEx', 'nEy', 'nEz', 'vnEx', 'vnEy', 'vnEz', '_freq',
-                     '_sval', '_smu0']
+        attr_list = ['nEx', 'nEy', 'nEz', 'vnEx', 'vnEy', 'vnEz', '_freq']
         for attr in attr_list:
             new_state += (getattr(self, attr),)
 
@@ -193,8 +180,7 @@ class Field(np.ndarray):
         => https://stackoverflow.com/a/26599346
         """
         # Set the necessary attributes (in reverse order).
-        attr_list = ['nEx', 'nEy', 'nEz', 'vnEx', 'vnEy', 'vnEz', '_freq',
-                     '_sval', '_smu0']
+        attr_list = ['nEx', 'nEy', 'nEz', 'vnEx', 'vnEy', 'vnEz', '_freq']
         attr_list.reverse()
         for i, name in enumerate(attr_list):
             i += 1  # We need it 1..#attr instead of 0..#attr-1.
@@ -205,7 +191,7 @@ class Field(np.ndarray):
 
     def copy(self):
         """Return a copy of the Field."""
-        return Field.from_dict(self.to_dict(True))
+        return self.from_dict(self.to_dict(True))
 
     def to_dict(self, copy=False):
         """Store the necessary information of the Field in a dict."""
@@ -335,11 +321,27 @@ class Field(np.ndarray):
     @property
     def smu0(self):
         """Return s*mu_0; mu_0 = Magn. permeability of free space [H/m]."""
+        if getattr(self, '_smu0', None) is None:
+            if self.sval is not None:
+                self._smu0 = self.sval*mu_0
+            else:
+                self._smu0 = None
+
         return self._smu0
 
     @property
     def sval(self):
         """Return s; s=iw in frequency domain; s=freq in Laplace domain."""
+
+        if getattr(self, '_sval', None) is None:
+            if self._freq is not None:
+                if self._freq < 0:  # Laplace domain; s.
+                    self._sval = np.array(self._freq)
+                else:  # Frequency domain; s = iw = 2i*pi*f.
+                    self._sval = np.array(-2j*np.pi*self._freq)
+            else:
+                self._sval = None
+
         return self._sval
 
     @property
@@ -380,7 +382,7 @@ class SourceField(Field):
     Parameters
     ----------
 
-    fx_or_grid : :class:`TensorMesh` or ndarray
+    fx_or_grid : :class:`emg3d.meshes.TensorMesh` or ndarray
         Either a TensorMesh instance or an ndarray of shape grid.nEx or
         grid.vnEx. See explanations above. Only mandatory parameter; if the
         only one provided, it will initiate a zero-field of `dtype`.
@@ -427,53 +429,6 @@ class SourceField(Field):
         return super().__new__(cls, fx_or_grid, fy_or_field=fy_or_field,
                                fz=fz, dtype=dtype, freq=freq)
 
-    def copy(self):
-        """Return a copy of the SourceField."""
-        return SourceField.from_dict(self.to_dict(True))
-
-    @classmethod
-    def from_dict(cls, inp):
-        """Convert dictionary into :class:`SourceField` instance.
-
-        Parameters
-        ----------
-        inp : dict
-            Dictionary as obtained from :func:`SourceField.to_dict`.
-            The dictionary needs the keys `field`, `freq`, `vnEx`, `vnEy`, and
-            `vnEz`.
-
-        Returns
-        -------
-        obj : :class:`SourceField` instance
-
-        """
-
-        # Create a dummy with the required attributes for the field instance.
-        class Grid:
-            pass
-
-        grid = Grid()
-
-        # Check and get the required keys from the input.
-        try:
-            field = inp['field']
-            freq = inp['freq']
-            grid.vnEx = inp['vnEx']
-            grid.vnEy = inp['vnEy']
-            grid.vnEz = inp['vnEz']
-        except KeyError as e:
-            print(f"* ERROR   :: Variable {e} missing in `inp`.")
-            raise
-
-        # Calculate missing info.
-        grid.nEx = np.prod(grid.vnEx)
-        grid.nEy = np.prod(grid.vnEy)
-        grid.nEz = np.prod(grid.vnEz)
-        grid.nE = grid.nEx + grid.nEy + grid.nEz
-
-        # Return Field instance.
-        return cls(grid, field, freq=freq)
-
     @property
     def vector(self):
         """Entire vector, 1D [vx, vy, vz]."""
@@ -517,7 +472,7 @@ def get_source_field(grid, src, freq, strength=0):
     Parameters
     ----------
     grid : TensorMesh
-        Model grid; a :class:`TensorMesh` instance.
+        Model grid; a :class:`emg3d.meshes.TensorMesh` instance.
 
     src : list of floats
         Source coordinates (m). There are two formats:
@@ -781,8 +736,8 @@ def get_receiver(grid, values, coordinates, method='cubic', extrapolate=False):
 
     Parameters
     ----------
-    grid : TensorMesh
-        Model grid; a :class:`TensorMesh` instance.
+    grid : :class:`emg3d.meshes.TensorMesh`
+        The model grid.
 
     values : ndarray
         Field instance, or a particular field (e.g. field.fx); Model
@@ -819,6 +774,7 @@ def get_receiver(grid, values, coordinates, method='cubic', extrapolate=False):
     See Also
     --------
     grid2grid : Interpolation of model parameters or fields to a new grid.
+    get_receiver_response : Get response for arbitrarily rotated receivers.
 
     """
     # If values is a Field instance, call it recursively for each field.
@@ -863,6 +819,93 @@ def get_receiver(grid, values, coordinates, method='cubic', extrapolate=False):
         return out
     else:
         return EMArray(out)
+
+
+def get_receiver_response(grid, field, rec):
+    """Return the field (response) at receiver coordinates.
+
+    Parameters
+    ----------
+    grid : :class:`emg3d.meshes.TensorMesh`
+        The model grid.
+
+    field : :class:`Field`
+        The electric or magnetic field.
+
+    rec : tuple (x, y, z, azimuth, dip)
+        Receiver coordinates and angles (m, °).
+
+        All values can either be a scalar or having the same length as number
+        of receivers.
+
+        Angles:
+
+        - azimuth (°): horizontal deviation from x-axis, anti-clockwise.
+        - dip (°): vertical deviation from xy-plane up-wards.
+
+
+    Returns
+    -------
+    responses : :class:`empymod.utils.EMArray`
+        Responses at receiver.
+
+
+    .. note::
+
+        Currently only implemented for point receivers, not for finite length
+        dipoles.
+
+
+    See Also
+    --------
+    get_receiver : Get values at coordinates (fields and models).
+
+    """
+
+    # Check receiver dimension.
+    if len(rec) != 5:
+        print("* ERROR   :: `rec` needs to be in the form "
+              "(x, y, z, azimuth, dip).\n             "
+              f"Length of provided `rec`: {len(rec)}.")
+        raise ValueError("Receiver error")
+
+    # Check field dimension to ensure it is not a particular field.
+    if field.field.ndim == 3:
+        print("* ERROR   :: `field` must be a `Field`-instance, not a\n"
+              "             particular field such as `field.fx`.")
+        raise ValueError("Field error")
+
+    # Get the vectors corresponding to input data.
+    if field.is_electric:
+        points = ((grid.vectorCCx, grid.vectorNy, grid.vectorNz),
+                  (grid.vectorNx, grid.vectorCCy, grid.vectorNz),
+                  (grid.vectorNx, grid.vectorNy, grid.vectorCCz))
+    else:
+        points = ((grid.vectorNx, grid.vectorCCy, grid.vectorCCz),
+                  (grid.vectorCCx, grid.vectorNy, grid.vectorCCz),
+                  (grid.vectorCCx, grid.vectorCCy, grid.vectorNz))
+
+    # Get azimuth and dip in radians.
+    azm = np.deg2rad(rec[3])
+    dip = np.deg2rad(rec[4])
+
+    # Get factors in the different directions.
+    factors = (np.cos(azm)*np.cos(dip),  # x
+               np.sin(azm)*np.cos(dip),  # y
+               np.sin(dip))  # z
+
+    # Pre-allocate the response.
+    resp = np.zeros(max([np.atleast_1d(x).size for x in rec]),
+                    dtype=field.dtype)
+
+    # Add the required responses.
+    inp = {'method': 'cubic', 'fill_value': 0.0, 'mode': 'constant'}
+    for i, ff in enumerate((field.fx, field.fy, field.fz)):
+        if np.any(abs(factors[i]) > 1e-10):
+            resp += factors[i]*maps.interp3d(points[i], ff, rec[:3], **inp)
+
+    # Return response.
+    return EMArray(resp)
 
 
 def get_h_field(grid, model, field):