+++ CHANGELOG.rst
@@ -5,11 +5,18 @@ Changelog
 *latest*
 --------
 
+- `fields`: New function `get_receiver_response`, which returns the response
+  for arbitrarily rotated receivers.
 - `io`: New `backend='json'`, writes to a hierarchical, plain json file.
   Changed backend name to actual file extension: `npz` (instead of `numpy`);
   `h5` (instead of `h5py`), and new `json`. (`numpy` and `h5py` are
   deprecated.)
 - Added `__eq__` to `models.TensorMesh` to compare meshes.
+- Improvements to `Field` and `SourceField`:
+
+  - `_sval` and `_smu0` not stored any longer, derived from `_freq`.
+  - `SourceField` is now using the `copy()` and `from_dict()` from its parents
+    class `Field`.
 
 
 *v0.11.0* : Refactor